+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 3.3.0-dev
+
+* Add `??` null-aware operator.
+* Add `..` cascade assignment operator.
+* Add `part` directive.
+
 ## 3.2.2
 
 * Require minimum Dart SDK of `2.6.0`.